Move EntriesManager & EntriesModel to separate file
